# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: 'OnesBlog API Documentation'
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://ones-blog-api.test'
paths:
...
    '/api/locations/{location}':
        get:
            summary: 'Display the specified location.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200, when location displayed.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: { user_id: '{location-user-id}', city_area_id: '{location-city-area-id}', category_id: '{location-category-id}', name: '{location-name}', address: '{location-address}', phone: '{location-phone}', avgScore: '{location-avgScore}', introduction: '{location-introduction}', active: '{location-active}', images: '{location-images}' }
                                properties:
                                    data: { type: object, example: { user_id: '{location-user-id}', city_area_id: '{location-city-area-id}', category_id: '{location-category-id}', name: '{location-name}', address: '{location-address}', phone: '{location-phone}', avgScore: '{location-avgScore}', introduction: '{location-introduction}', active: '{location-active}', images: '{location-images}' } }
                404:
                    description: '404, when location not found or unverified.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'Location(ID:{location-id}) is not found.'
                                properties:
                                    data: { type: string, example: 'Location(ID:{location-id}) is not found.' }
            tags:
                - '02. Locations'
            security: []
        put:
            summary: 'Update the specified location in storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Authorization
                    description: ''
                    example: 'Bearer {personal-access-token}'
                    schema:
                        type: string
            responses:
                200:
                    description: '200, when location displayed.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: { user_id: '{location-user-id}', city_area_id: '{location-city-area-id}', category_id: '{location-category-id}', name: '{location-name}', address: '{location-address}', phone: '{location-phone}', avgScore: '{location-avgScore}', introduction: '{location-introduction}', images: '{location-images}' }
                                properties:
                                    data: { type: object, example: { user_id: '{location-user-id}', city_area_id: '{location-city-area-id}', category_id: '{location-category-id}', name: '{location-name}', address: '{location-address}', phone: '{location-phone}', avgScore: '{location-avgScore}', introduction: '{location-introduction}', images: '{location-images}' } }
                401:
                    description: '401, without personal access token.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: Unauthenticated.
                                properties:
                                    data: { type: string, example: Unauthenticated. }
                403:
                    description: '403, when location updated by wrong user.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'This action is unauthorized.'
                                properties:
                                    data: { type: string, example: 'This action is unauthorized.' }
                404:
                    description: '404, when location not found.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'Location(ID:{location-id}) is not found.'
                                properties:
                                    data: { type: string, example: 'Location(ID:{location-id}) is not found.' }
                422:
                    description: '422, when any validation failed.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: { city_area_id: ['{validation-error-message}'], category_id: ['{validation-error-message}'] }
                                properties:
                                    data: { type: object, example: { city_area_id: ['{validation-error-message}'], category_id: ['{validation-error-message}'] } }
            tags:
                - '02. Locations'
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                city_area_id:
                                    type: integer
                                    description: 'The city area of the location.'
                                    example: 153
                                category_id:
                                    type: integer
                                    description: 'The category of the location.'
                                    example: 2
                                name:
                                    type: string
                                    description: 'The name of the location.'
                                    example: 統一娛樂
                                address:
                                    type: string
                                    description: 'The address of the location.'
                                    example: 豐裡二路180巷804弄601號49樓
                                phone:
                                    type: string
                                    description: 'The phone of the location.'
                                    example: '1335933680'
                                introduction:
                                    type: string
                                    description: 'The introduction of the location.'
                                    example: IntroductionTest
                                active:
                                    type: boolean
                                    description: 'The state of the location.'
                                    example: true
                                images:
                                    type: array
                                    description: 'The images of the location.'
                                    example: [null, null]
                                    items: { type: file }
                                _method:
                                    type: string
                                    description: 'Required if the <code><b>images</b></code> of the location are uploaded, must be <b>PUT</b> and request method must be <small class="badge badge-black">POST</small>.'
                                    example: PUT
                            required:
                                - city_area_id
                                - category_id
                                - name
                                - address
                                - phone
        delete:
            summary: 'Remove the specified location from storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Authorization
                    description: ''
                    example: 'Bearer {personal-access-token}'
                    schema:
                        type: string
            responses:
                200:
                    description: '200, when location deleted.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: Success
                                properties:
                                    data: { type: string, example: Success }
                401:
                    description: '401, without personal access token.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: Unauthenticated.
                                properties:
                                    data: { type: string, example: Unauthenticated. }
                403:
                    description: '403, when location deleted by wrong user.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'This action is unauthorized.'
                                properties:
                                    data: { type: string, example: 'This action is unauthorized.' }
                404:
                    description: '404, when location not found.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'Location(ID:{location-id}) is not found.'
                                properties:
                                    data: { type: string, example: 'Location(ID:{location-id}) is not found.' }
            tags:
                - '02. Locations'
        parameters:
            -
                in: path
                name: location
                description: 'The id of the location.'
                example: 5
                required: true
                schema:
                    type: integer
...